requirements.txt
certifi==2020.6.20
chardet==3.0.4
flake8==3.8.3
idna==2.10
imgurpython==1.1.7
importlib-metadata==1.7.0
mccabe==0.6.1
mysql-connector-python==8.0.21
Pillow==7.2.0
praw==7.1.0
prawcore==1.5.0
protobuf==3.13.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
zipp==3.1.0
